+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **Major:** Add trajectory-player agent
 - agent: Add `start_mpc_balancer.sh` script to the repository
 - agent: Import MPC balancer from previous Pink balancer
+- config: Load user configuration from `~/.config/upkie`
 - control: Create `upkie.controllers` submodule
 - control: Import MPC balancer class from the Pink balancer
 - cpp: Add documentation to the moteus actuation interface